+++ lib/celluloid.rb
@@ -200,20 +200,6 @@ module Celluloid
       Supervisor.supervise_as(name, self, *args, &block)
     end
 
-    # Create a new pool of workers. Accepts the following options:
-    #
-    # * size: how many workers to create. Default is worker per CPU core
-    # * args: array of arguments to pass when creating a worker
-    #
-    def pool(options = {})
-      PoolManager.new(self, options)
-    end
-
-    # Same as pool, but links to the pool manager
-    def pool_link(options = {})
-      PoolManager.new_link(self, options)
-    end
-
     # Run an actor in the foreground
     def run(*args, &block)
       Actor.join(new(*args, &block))
@@ -509,7 +495,6 @@ require 'celluloid/actor'
 require 'celluloid/cell'
 require 'celluloid/future'
 require 'celluloid/actor_system'
-require 'celluloid/pool_manager'
 require 'celluloid/supervision_group'
 require 'celluloid/supervisor'
 require 'celluloid/notifications'